--- api/src/sockets/socketDictionary.js
@@ -1,2 +1,0 @@
-let socketDictionary = {}
-module.exports = {socketDictionary};
